+++ d4np/d4np_c.h
@@ -43,6 +43,13 @@
 #include "d4np/io/file.h"
 #include "d4np/io/path.h"
 
+/* System utilities & diagnostics */
+#include "d4np/sys/clock.h"
+#include "d4np/sys/error_context.h"
+#include "d4np/sys/hash.h"
+#include "d4np/sys/log.h"
+#include "d4np/sys/uuid.h"
+
 /* Strings & parsing */
 #include "d4np/str/parse.h"
 #include "d4np/str/str_view.h"