+++ tests/tests.h
@@ -26,5 +26,10 @@ void suite_atomic_queue(void);
 void suite_thread_pool(void);
 void suite_file(void);
 void suite_path(void);
+void suite_hash(void);
+void suite_clock(void);
+void suite_uuid(void);
+void suite_error_context(void);
+void suite_log(void);
 
 #endif /* D4NP_TESTS_H */